requirements.txt
google-generativeai==0.3.2
playwright==1.40.0
flask==3.0.0
flask-cors==4.0.0
slack-sdk==3.23.0
PyGithub==2.1.1
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0
cryptography==41.0.7
sqlalchemy==2.0.23
apscheduler==3.10.4

